Relocate network utils tests to tests/utils/test_net (#585)
