+++ tests/test_init.py
@@ -4,7 +4,6 @@
 
 """ Unit tests for zeroconf.py """
 
-import errno
 import logging
 import socket
 import time
@@ -502,21 +501,6 @@ def test_ptr_optimization():
     zc.close()
 
 
-@pytest.mark.parametrize(
-    "errno,expected_result",
-    [(errno.EADDRINUSE, False), (errno.EADDRNOTAVAIL, False), (errno.EINVAL, False), (0, True)],
-)
-def test_add_multicast_member_socket_errors(errno, expected_result):
-    """Test we handle socket errors when adding multicast members."""
-    if errno:
-        setsockopt_mock = unittest.mock.Mock(side_effect=OSError(errno, "Error: {}".format(errno)))
-    else:
-        setsockopt_mock = unittest.mock.Mock()
-    fileno_mock = unittest.mock.PropertyMock(return_value=10)
-    socket_mock = unittest.mock.Mock(setsockopt=setsockopt_mock, fileno=fileno_mock)
-    assert r.add_multicast_member(socket_mock, "0.0.0.0") == expected_result
-
-
 def test_notify_listeners():
     """Test adding and removing notify listeners."""
     # instantiate a zeroconf instance
@@ -553,11 +537,3 @@ def test_notify_listeners():
     assert not notify_called
 
     zc.close()
-
-
-def test_autodetect_ip_version():
-    """Tests for auto detecting IPVersion based on interface ips."""
-    assert r.autodetect_ip_version(["1.3.4.5"]) is r.IPVersion.V4Only
-    assert r.autodetect_ip_version([]) is r.IPVersion.V4Only
-    assert r.autodetect_ip_version(["::1", "1.2.3.4"]) is r.IPVersion.All
-    assert r.autodetect_ip_version(["::1"]) is r.IPVersion.V6Only
