+++ tests/utils/test_net.py
@@ -5,10 +5,13 @@
 """Unit tests for zeroconf._utils.net."""
 from unittest.mock import Mock, patch
 
+import errno
 import ifaddr
 import pytest
+import unittest
 
 from zeroconf._utils import net as netutils
+import zeroconf as r
 
 
 def _generate_mock_adapters():
@@ -56,3 +59,26 @@ def test_ip6_addresses_to_indexes():
     interfaces = ['2001:db8::']
     with patch("zeroconf._utils.net.ifaddr.get_adapters", return_value=_generate_mock_adapters()):
         assert netutils.ip6_addresses_to_indexes(interfaces) == [(('2001:db8::', 1, 1), 1)]
+
+
+@pytest.mark.parametrize(
+    "errno,expected_result",
+    [(errno.EADDRINUSE, False), (errno.EADDRNOTAVAIL, False), (errno.EINVAL, False), (0, True)],
+)
+def test_add_multicast_member_socket_errors(errno, expected_result):
+    """Test we handle socket errors when adding multicast members."""
+    if errno:
+        setsockopt_mock = unittest.mock.Mock(side_effect=OSError(errno, "Error: {}".format(errno)))
+    else:
+        setsockopt_mock = unittest.mock.Mock()
+    fileno_mock = unittest.mock.PropertyMock(return_value=10)
+    socket_mock = unittest.mock.Mock(setsockopt=setsockopt_mock, fileno=fileno_mock)
+    assert r.add_multicast_member(socket_mock, "0.0.0.0") == expected_result
+
+
+def test_autodetect_ip_version():
+    """Tests for auto detecting IPVersion based on interface ips."""
+    assert r.autodetect_ip_version(["1.3.4.5"]) is r.IPVersion.V4Only
+    assert r.autodetect_ip_version([]) is r.IPVersion.V4Only
+    assert r.autodetect_ip_version(["::1", "1.2.3.4"]) is r.IPVersion.All
+    assert r.autodetect_ip_version(["::1"]) is r.IPVersion.V6Only
